TST: merge unused _test_ci.py into test_confidence.py
