+++ tests/test_confidence.py
@@ -78,6 +78,50 @@ def test_confidence2():
 
     lmfit.printfuncs.report_ci(ci)
 
-if __name__ == '__main__':
-    test_confidence1()
-    test_confidence2()
+
+
+def test_ci_with_trace():
+    np.random.seed(1)
+    p_true = lmfit.Parameters()
+    p_true.add('amp', value=14.0)
+    p_true.add('period', value=5.33)
+    p_true.add('shift', value=0.123)
+    p_true.add('decay', value=0.010)
+
+    def residual(pars, x, data=None):
+        amp = pars['amp']
+        per = pars['period']
+        shift = pars['shift']
+        decay = pars['decay']
+
+        if abs(shift) > np.pi / 2:
+            shift = shift - np.sign(shift) * np.pi
+        model = amp * np.sin(shift + x / per) * np.exp(-x * x * decay * decay)
+        if data is None:
+            return model
+        return model - data
+
+    n = 2500
+    xmin = 0.
+    xmax = 250.0
+    noise = np.random.normal(scale=0.7215, size=n)
+    x = np.linspace(xmin, xmax, n)
+    data = residual(p_true, x) + noise
+
+    fit_params = lmfit.Parameters()
+    fit_params.add('amp', value=13.0)
+    fit_params.add('period', value=4)
+    fit_params.add('shift', value=0.1)
+    fit_params.add('decay', value=0.02)
+
+    mini = lmfit.Minimizer(residual, fit_params, fcn_args=(x, data))
+    out = mini.minimize()
+    fit = residual(fit_params, x)
+
+    ci, tr = lmfit.conf_interval(mini, out, sigmas=[0.674], trace=True)
+    for p in out.params:
+        diff1 = ci[p][1][1] - ci[p][0][1]
+        diff2 = ci[p][2][1] - ci[p][1][1]
+        stderr = out.params[p].stderr
+        assert(abs(diff1 - stderr) / stderr < 0.05)
+        assert(abs(diff2 - stderr) / stderr < 0.05)
